requirements.txt
beautifulsoup4 == 4.9.3
hydra_core == 1.1.1
multiprocess == 0.70.11.1
numpy == 1.18.5
omegaconf == 2.1.1
pytorch_lightning == 1.5.1
pytorch_nlp == 0.5.0
razdel == 0.5.0
torch == 1.7.0
torchmetrics == 0.7.1
transformers == 4.5.0
typing_extensions == 4.0.1
